update filenames and tests
