+++ src/components/publishing/__test__/imageset_preview_classic.test.tsx
@@ -1,0 +1,10 @@
+import ImagesetPreviewClassic from "../imageset_preview_classic"
+import { Images } from "./fixtures"
+
+import * as React from "react"
+import * as renderer from "react-test-renderer"
+
+it("renders properly", () => {
+  const imageset = renderer.create(<ImagesetPreviewClassic images={Images} />).toJSON()
+  expect(imageset).toMatchSnapshot()
+})
