+++ src/components/publishing/index.tsx
@@ -2,10 +2,12 @@ import Artwork from "./artwork"
 import FeatureHeader from "./feature_header"
 import Image from "./image"
 import ImagesetPreview from "./imageset_preview"
+import ImagesetPreviewClassic from "./imageset_preview_classic"
 
 export default {
   Artwork,
   FeatureHeader,
   Image,
   ImagesetPreview,
+  ImagesetPreviewClassic,
 }
